requirements.txt
django
djangorestframework
python-dotenv
dnspython
python-whois
requests

